code/requirements.txt
alembic==1.7.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopage==0.5.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cffi==1.15.0
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorlog==6.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastdtw==0.3.4
fastjsonschema==2.15.3
fonttools==4.32.0
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.20.3
openpyxl==3.0.7
optuna==2.10.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prettytable==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
scikit-learn==1.0.2
scipy==1.6.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
slack-sdk==3.19.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
stack-data==0.2.0
stevedore==3.5.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.61.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xlrd==2.0.1
zipp==3.8.0